refactor(all): standardize architecture, add topology tools and overhaul documentation (#7)

+++ src/model/torus.cpp
@@ -2,13 +2,15 @@
 
 #include <vector>
 #include <cmath>
+#include <cstddef> // offsetof
 #include <glm/gtc/constants.hpp>
 
 namespace model {
 
+    // Internal helper struct for interleaving vertex data
     struct Vertex {
         float position[3];
-        float normal[3];   // currently not used (only ambient ilumination in the shader)
+        float normal[3];
         float uv[2];
     };
 
@@ -17,43 +19,52 @@ namespace model {
         const int vertexRows = rows + 1;
         constexpr float TWO_PI = glm::two_pi<float>();
 
-        std::vector<Vertex> vertices(static_cast<std::size_t>(vertexCols) * static_cast<std::size_t>(vertexRows));
-        auto vertexIndex = [&](int i, int j) -> int { return j * vertexCols + i; };
+        std::vector<Vertex> vertices(static_cast<size_t>(vertexCols) * vertexRows);
+
+        // Helper lambda to map 2D grid coordinates to linear index
+        auto vertexIndex = [&](int i, int j) -> int {
+            return j * vertexCols + i;
+            };
 
         // Generate vertices
         for (int j = 0; j < vertexRows; ++j) {
             const float v = static_cast<float>(j) / static_cast<float>(rows);
-            const float angMinor = (0.5f - v) * TWO_PI;  // minor angle (tube)
+            const float angMinor = (0.5f - v) * TWO_PI; // minor angle (tube)
             const float cosMinor = std::cos(angMinor);
             const float sinMinor = std::sin(angMinor);
 
             for (int i = 0; i < vertexCols; ++i) {
                 const float u = static_cast<float>(i) / static_cast<float>(columns);
-                const float angMajor = u * TWO_PI;        // major angle (ring)
+                const float angMajor = u * TWO_PI; // major angle (ring)
                 const float cosMajor = std::cos(angMajor);
                 const float sinMajor = std::sin(angMajor);
 
                 const float ring = outerRadius + innerRadius * cosMinor;
 
                 Vertex vertex{};
+
+                // Position
                 vertex.position[0] = ring * cosMajor;
                 vertex.position[1] = ring * sinMajor;
                 vertex.position[2] = innerRadius * sinMinor;
 
+                // Normal (analytical)
                 vertex.normal[0] = cosMajor * cosMinor;
                 vertex.normal[1] = sinMajor * cosMinor;
                 vertex.normal[2] = sinMinor;
 
-                vertex.uv[0] = 1.0f - u;  // mirror U so grid-right maps to torus-right
+                // UVs
+                vertex.uv[0] = 1.0f - u; // mirror U so grid-right maps to torus-right
                 vertex.uv[1] = v;
 
-                vertices[static_cast<std::size_t>(vertexIndex(i, j))] = vertex;
+                vertices[static_cast<size_t>(vertexIndex(i, j))] = vertex;
             }
         }
 
         // Build triangle indices (two triangles per quad)
         std::vector<unsigned int> indices;
-        indices.reserve(static_cast<std::size_t>(columns) * static_cast<std::size_t>(rows) * 6);
+        indices.reserve(static_cast<size_t>(columns) * rows * 6);
+
         for (int j = 0; j < rows; ++j) {
             for (int i = 0; i < columns; ++i) {
                 const unsigned int i0 = static_cast<unsigned int>(vertexIndex(i, j));
@@ -61,10 +72,12 @@ namespace model {
                 const unsigned int i2 = static_cast<unsigned int>(vertexIndex(i, j + 1));
                 const unsigned int i3 = static_cast<unsigned int>(vertexIndex(i + 1, j + 1));
 
+                // Triangle 1
                 indices.push_back(i0);
                 indices.push_back(i2);
                 indices.push_back(i1);
 
+                // Triangle 2
                 indices.push_back(i1);
                 indices.push_back(i2);
                 indices.push_back(i3);
@@ -92,10 +105,16 @@ namespace model {
             GL_STATIC_DRAW);
 
         const GLsizei stride = static_cast<GLsizei>(sizeof(Vertex));
+
+        // Attribute 0: Position
         glEnableVertexAttribArray(0);
         glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, stride, (void*)offsetof(Vertex, position));
+
+        // Attribute 1: Normal
         glEnableVertexAttribArray(1);
         glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, stride, (void*)offsetof(Vertex, normal));
+
+        // Attribute 2: UV
         glEnableVertexAttribArray(2);
         glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, stride, (void*)offsetof(Vertex, uv));
 
@@ -109,6 +128,7 @@ namespace model {
         if (mesh.ebo_) glDeleteBuffers(1, &mesh.ebo_);
         if (mesh.vbo_) glDeleteBuffers(1, &mesh.vbo_);
         if (mesh.vao_) glDeleteVertexArrays(1, &mesh.vao_);
+
         mesh = {}; // reset handles to 0
     }
